:recycle:refactor: 상품 댓글 TS 마이그레이션
