+++ src/constants/comment.ts
@@ -1,3 +1,13 @@
+export interface CommentWriter {
+  image: string;
+  nickname: string;
+  id: number;
+}
+
 export default interface Comment {
-  content: "string";
+  writer: CommentWriter;
+  updatedAt: string;
+  createdAt: string;
+  content: string;
+  id: number;
 }
